completed client refactoring
