+++ api/users/client-router.js
@@ -5,8 +5,8 @@ const { checkIfPasswordExists, checkIfEmailExists } = require('../auth/auth-midd
 const Trader = require('./trader-model');
 const axios = require('axios');
 const User = require('./user-model');
-const bcrypt = require('bcrypt');
-const { processClientBuyBitcoinOrder } = require("./client-middleware");
+
+const { processClientBuyBitcoinOrder, processClientSellBitcoinOrder } = require("./client-middleware");
 
 // /latest: endpoint that retrieves the latest bitcoin price
 router.get('/latest', async (req, res, next) => {
@@ -91,212 +91,61 @@ router.post('/buy-bitcoin', checkIfEmailExists, checkIfPasswordExists, processCl
                     amount: balance
                 })
         }
-
     }
     catch (err) {
         res.status(500)
             .json(`Server Error: ${err.message}`)
     }
 
-
-
 })
 
-// path to sell bitcoin 
-router.post('/sell-bitcoin', async (req, res, next) => {
+// /sell-bitcoin: endpoint that processes the selling of bitcoin of a client.
+router.post('/sell-bitcoin', checkIfEmailExists, checkIfPasswordExists, processClientSellBitcoinOrder, async (req, res, next) => {
     try {
-        const decoded = jwtDecode(req.headers.authorization)
-        const order = req.body
+        // get the balance
+        const balance = req.balance;
 
-        // retrieve client info
-        const client = await Client.retrieveClientInfo(decoded.email)
+        // retrieve is middleware operations were successful
+        const isOpSuccessful = req.isOpSuccessful;
 
-        // validate identity
-        const { email, password } = req.body
-        const user = await User.findByEmail(email)
-        const encryption = bcrypt.compareSync(password, user.password)
-        if (!(user && encryption)) {
-            res.status(400)
-                .json('Invalid credentials. Identity not confirmed')
-        }
-
-        // retrieve current bitcoin balance and check to see if sale can be made
-
-        if (client.Bitcoin_balance < order.Bitcoin_balance ||
-            isNaN(client.Bitcoin_balance) || client.Bitcoin_balance < 0) {
-            res.status(401)
-                .json('Client does not posses enough bitcoin to perform sell')
-
-        }
-        else {
-
-            // update balance of bitcoin and balance of usd
-            let updatedBitcoin = client.Bitcoin_balance - order.Bitcoin_balance
-            let updatedBalance = client.USD_balance +
-                (order.Bitcoin_balance *
-                    order.Bitcoin_price)
-            let commissionPay = 0.00
-            let commissionPayUSD;
-            let commissionPayBitcoin;
-
-            // update balance or bitcoin balance based on selection
-            // of how client wants to pay commission
-            if (order.comm_type === 'USD') {
-
-                if (client.mem_level === 'Silver') {
-                    commissionPay = (order.Bitcoin_balance *
-                        order.Bitcoin_price) * 0.1
-                    if (client.USD_balance < commissionPay) {
-                        res.status(401)
-                            .json('Client does not possess enough fiat USD to pay commission')
-                    }
-                    else {
-                        updatedBalance -= commissionPay
-                        commissionPayUSD = commissionPay
-
-                    }
-
-                }
-                else if (client.mem_level === 'Gold') {
-
-                    commissionPay = (order.Bitcoin_balance *
-                        order.Bitcoin_price) * 0.05
-                    if (client.USD_balance < commissionPay) {
-                        res.status(401)
-                            .json('Client does not possess enough fiat USD to pay commission')
-                    }
-                    else {
-                        updatedBalance -= commissionPay
-                        commissionPayUSD = commissionPay
-
-                    }
-
-                }
-
-            }
-            else if (order.comm_type === 'Bitcoin') {
-                if (client.mem_level === 'Silver') {
-                    commissionPay = (order.Bitcoin_balance) * 0.1
-                    if (client.Bitcoin_balance < commissionPay) {
-                        res.status(401)
-                            .json('Client does not possess enough bitcoin to pay commission')
-                    }
-                    else {
-                        updatedBitcoin -= commissionPay
-                        commissionPayBitcoin = commissionPay
-
-                    }
-
-                }
-                else if (client.mem_level === 'Gold') {
-
-                    commissionPay = (order.Bitcoin_balance) * 0.05
-                    if (client.Bitcoin_balance < commissionPay) {
-                        res.status(401)
-                            .json('Client does not possess enough bitcoin to pay commission')
-                    }
-                    else {
-                        updatedBitcoin -= commissionPay
-                        commissionPayBitcoin = commissionPay
-
-                    }
-
-                }
-            }
-
-
-            // insert balance and bitcoin into tables to update
-            const updateBitcoin = await Client.updateBitcoinWallet(decoded.email,
-                updatedBitcoin)
-            const updateUSD = await Client.updateUSDBalance(decoded.email,
-                updatedBalance)
-            const date = new Date()
-            const formattedDate = `${date.getFullYear()}` + '-' + `${date.getMonth()}` + '-' + `${date.getDate()}`
-
-            // create object for record of order
-            const orderCreds = {
-                client_id: client.client_id,
-                date: formattedDate,
-                comm_paid: commissionPay,
-                comm_type: order.comm_type,
-                Bitcoin_balance: order.Bitcoin_balance,
-                isCancelled: false
-
-            }
-            // increment trades by 1 and update number of client trades
-            const incrementedTrades = (client.num_trades + 1)
-            const updateNumTrades = await Client.updateNumTrades(client.email, incrementedTrades)
-
-            // Update trader balance--based on comm_type
-
-            let updateUSDBalanceOfTrader;
-            let updateBitcoinBalanceOfTrader;
-
-            // insert order into order table
-            const addOrder = await Client.addOrder(orderCreds)
-
-            if (order.comm_type === 'USD') {
-                updateUSDBalanceOfTrader = await Trader.updateUSDBalanceOfTrader(client.trader_id, commissionPayUSD)
-
-
-                if (addOrder &&
-                    client &&
-                    updateBitcoin &&
-                    updateUSD &&
-                    updateNumTrades &&
-                    updateUSDBalanceOfTrader) {
-                    res.status(201)
-                        .json({
-                            message: 'successfully sold bitcoin',
-                            amount: order.Bitcoin_balance
-                        })
-                }
-            }
-            else if (order.comm_type === 'Bitcoin') {
-                updateBitcoinBalanceOfTrader = await Trader.updateBitcoinBalanceOfTrader(client.trader_id, commissionPayBitcoin)
-
-
-                if (addOrder &&
-                    client &&
-                    updateBitcoin &&
-                    updateUSD &&
-                    updateNumTrades &&
-                    updateBitcoinBalanceOfTrader) {
-                    res.status(201)
-                        .json({
-                            message: 'successfully sold bitcoin',
-                            amount: order.Bitcoin_balance
-                        })
-                }
-            }
-
-
+        // check if op successful.
+        if (isOpSuccessful) {
+            // op successful send success response.
+            res.status(201)
+                .json({
+                    message: 'successfully sold bitcoin',
+                    amount: balance
+                })
         }
     }
     catch (err) {
+        // internal server error, send failure response.
         res.status(500)
             .json(`Server Error: ${err.message}`)
-
     }
-
-
-
 })
 
 
-// path to retrieve bitcoin wallet of client
-router.get('/BitcoinWallet', async (req, res, next) => {
+// /bitcoin-wallet: path to retrieve bitcoin wallet of client
+router.get('/bitcoin-wallet', async (req, res, next) => {
     try {
-        const decoded = jwtDecode(req.headers.authorization)
-        const wallet = await Client.retrieveClientInfo(decoded.email)
+        // decode token from auth header
+        const decoded = jwtDecode(req.headers.authorization);
 
+        // retrieve the wallet from db.
+        const wallet = await Client.retrieveClientInfo(decoded.email);
+
+        //check if wallet was retrieved.
         if (wallet) {
+            // wallet was retrieved, send success response with wallet.
             res.status(200)
                 .json(wallet.Bitcoin_balance)
         }
 
     }
     catch (err) {
+
+        // internal server error, send failure response.
         res.status(500)
             .json(`Server Error: ${err.message}`)
     }
@@ -306,23 +155,29 @@ router.get('/BitcoinWallet', async (req, res, next) => {
 })
 
 
-// path to transfer money to trader
-router.post('/TransferMoney', async (req, res, next) => {
+// /transfer-money: path to transfer money to trader
+router.post('/transfer-money', async (req, res, next) => {
     try {
 
-        // check the usd balance first
-        const decoded = jwtDecode(req.headers.authorization)
-        const client = await Client.retrieveClientInfo(decoded.email)
-        const transfer = req.body
+        // decode token from auth header
+        const decoded = jwtDecode(req.headers.authorization);
 
+        // retrieve client info from db using email
+        const client = await Client.retrieveClientInfo(decoded.email);
+
+        // retrieve transfer amount from req body
+        const transfer = req.body;
+
+        // check if usd is 0.
         if (client.USD_balance <= 0) {
+            // dont have enough usd, send failure response.
             res.status(401)
                 .json('You do not have enough usd in your account')
         }
 
-        // retrieve client id, trader id
 
 
+        // get the current date and format it for db insertion
         const date = new Date()
         const formattedDate = `${date.getFullYear()}` + '-' + `${date.getMonth()}` + '-' + `${date.getDate()}`
 
@@ -343,16 +198,20 @@ router.post('/TransferMoney', async (req, res, next) => {
         // update the usd balance of the client
         const reducedBalance = client.USD_balance - transfer.amount_paid
 
+        // update the usd balance in db and save result
         const updateUSDBalance = await Client.updateUSDBalance(decoded.email, reducedBalance)
 
         // update the transfer account of the trader
         const updateTransferAccount = await Trader.updateTransferAccountById(client.trader_id, transfer.amount_paid)
 
 
+        // check if operations were successful
         if (client &&
             transferMoney &&
             updateUSDBalance &&
             updateTransferAccount) {
+
+            // operations succeeded, send success response.
             res.status(201)
                 .json('Successfully transferred payment to trader')
 
@@ -362,6 +221,7 @@ router.post('/TransferMoney', async (req, res, next) => {
 
     }
     catch (err) {
+        // internal server error, send failure response.
         res.status(500)
             .json(`Server Error: ${err.message}`)
     }
